オーバーライド処理を追加
