+++ src/feature/Hand/index.tsx
@@ -1,19 +1,25 @@
 import { HandView } from "@/component/ui/HandView"
+import { useWebSocketGame } from "@/hooks/game";
 import { useSystemContext } from "@/hooks/system/hooks";
 import { ICard } from "@/submodule/suit/types";
-import { useDndMonitor, DragStartEvent } from "@dnd-kit/core";
+import { useDndMonitor, DragStartEvent, DragEndEvent } from "@dnd-kit/core";
 
 interface HandAreaProps {
   hand?: ICard[]
 }
 
 export const HandArea = ({ hand }: HandAreaProps) => {
-  const { setActiveCard } = useSystemContext();
+  const { activeCard, setActiveCard } = useSystemContext();
+  const { override } = useWebSocketGame()
   useDndMonitor({
     onDragStart(e: DragStartEvent) {
       setActiveCard(e.active)
     },
-    onDragEnd() {
+    onDragEnd(e: DragEndEvent) {
+      const { over } = e;
+      if (over != null && activeCard?.id) {
+        override({ target: activeCard?.id as string, parent: over.id as string })
+      }
       setActiveCard(undefined)
     }
   })
